Worked on saving Teacher, Returning total workload
Created controller in Teacher entity, made changes in ID and used sequenceGenerator for it. Also, defined two methods in TeacherRepository and used it in service.Overriding the teacher totalworkhours method and returning the value in double data type. Used Java-Optionals and streams to calculate the business logic in service implementation layer.