Implemented controller of saving routine info and total workload.
Created required Mapping in GroupController. Defined required methods for creating a group in service layer. Implemented and override the methods for creating a new group and also implemented a method to return TotalGroupWorkHours which is extracted from list of routines saved in the database, implemented required logic and required exception is handled for now.